Revert "Separate module per component"
This reverts commit a71e70ff108f069899a732c7f8e72915db793350.
