--- packages/kiwi-react/src/bricks/dropdown-menu/DropdownMenuItem.tsx
@@ -1,20 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-import * as React from "react";
-import * as Ariakit from "@ariakit/react";
-import { ListItem } from "../ListItem.js";
-
-export const DropdownMenuItem = React.forwardRef<
-	React.ElementRef<typeof Ariakit.MenuItem>,
-	Ariakit.MenuItemProps
->((props, forwardedRef) => {
-	return (
-		<Ariakit.MenuItem
-			{...props}
-			render={<ListItem render={props.render} />}
-			ref={forwardedRef}
-		/>
-	);
-});
